requirements.txt
torch>=2.0.0
torchvision
basicsr @ file://./BasicSR
realesrgan
streamlit
Pillow
numpy

